add entities extractors; remove texture heap; fix bug with shape extraction

+++ include/functions.h
@@ -830,7 +830,7 @@ void init_trigger_list(void);
 void partner_init_after_battle(s32 arg0);
 void load_map_script_lib(void);
 void remove_item_entity_by_index(s32 index);
-void set_entity_commandlist(Entity* entity, s32* entityScript);
+void set_entity_commandlist(Entity* entity, intptr_t* entityScript);
 bool is_player_dismounted(void);
 void func_800EF300(void);
 void func_800EF314(void);
